- MOVE Static FL assets to their respective folders - All the static federated learning assets were moved to grid.app.main.sfl path - MOVE Satic FL websocket events (grid.app.main.events.federated) to its respective path (grid.app.main.events.sfl) - MOVE Static FL HTTP routes to its respective path (grid.app.main.routes.sfl.federated). - FIX import paths in sfl assets structures and schemas. - FIX import paths in test
